Create README.md
Tasks for Islamabad Quarter 1 
Class 7

Task 1: 
Write a TypeScript program that does the following:
    1. Declare three variables englishMarks, mathsMarks, and scienceMarks.
    2. Use the prompt-sync library to get input from the user for each subject and assign them to respective variables.
    3. Calculate the average marks of the student using the formula: average = (englishMarks + mathsMarks + scienceMarks) / 3.
    4. Display the average marks to the user as output.



Task 2: 
Basic Calculator using inquirer
Install the inquirer package using npm.
Use the inquirer package to prompt the user to select an operation (addition, subtraction, multiplication, division) and enter two numbers.
Based on the selected operation, perform the corresponding mathematical operation and display the result to the user as output.



Task 3: 
To-Do List using inquirer
Install the inquirer package using npm.
Use the inquirer package to prompt the user to enter a task and add it to a list of tasks.
Display the list of tasks to the user and prompt them to mark a task as completed or delete a task from the list.
Display the updated list of tasks to the user after each operation.


Task 4: 
Finding Minimum and Maximum Sum of Integers in TypeScript
Create a TypeScript program that includes an array of five positive integers.
Use an algorithm to find the minimum and maximum sums that can be calculated by summing exactly four of the five integers in the array.
Display the respective minimum and maximum values to the user as output in a single line of two space-separated long integers.

arr=[1,3,5,7,9]
Example
The minimum sum is 1+3+5+7=16 and
the maximum sum is 3+5+7+9=24
The function prints 16 24 (Required output)



Challenge task
Task 5: 
Converting AM/PM Time to Military Time in TypeScript

Create a TypeScript program that takes a time in the hour AM/PM format as input from the user using the inquirer package. The program should then convert the time to military (24-hour) time.

Note: 12:00:00AM on a 12-hour clock is 00:00:00 on a 24-hour clock. 12:00:00PM on a 12-hour clock is 12:00:00 on a 24-hour clock.

Use the inquirer package to prompt the user to enter a time in the hour AM/PM format.
Convert the input time to military (24-hour) time using an algorithm.
Display the converted time to the user as output.